hr update

+++ HrServiceCenterWeb/Models/CounterBO.cs
@@ -9,6 +9,6 @@ namespace HrServiceCenterWeb.Models
     {
         public string DataAxis { get; set; }
 
-        public int Data { get; set; }
+        public decimal Data { get; set; }
     }
 }
